emails_pythn/requirements.txt
IMAPClient==2.1.0
pyzmail36==1.0.4
six==1.14.0
